Update router function calls and remove unused dependencies
This commit mainly updates the calling of two functions in the router.js file to correct the method names. It also removes the `react-helmet` package from the list of dependencies since it is no longer being used in the project. In addition to this, a minor HTML visual enhancement was done for the 'Load More' button in the auction listing view.